Budget service: refactor directory structure
