+++ budget/src/app.ts
@@ -6,12 +6,7 @@ import cookieSession from 'cookie-session';
 import { currentUser } from '@bookkeeping/common';
 import { errorHandler, NotFoundError } from '@bookkeeping/common';
 
-import {
-  budgetIndex,
-  budgetRead,
-  budegetCreate,
-  budgetUpdate,
-} from './routes/budget';
+import { budgetIndex, budgetRead, budegetCreate, budgetUpdate } from './routes';
 
 const app = express();
 app.set('trust proxy', true);
